git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/experimental@3808 464221e2-0a0f-0410-983b-f8d57d210c17

+++ SourceCode/Framework/source/jcom.list.h
@@ -80,7 +80,7 @@ public:
 		ll.merge(m, comp);
 	}
 	/** @return the number of elements in the list */	
-	int size() { return ll.size(); }
+	long size() { return (long)ll.size(); }
 	/** @return true if the list is empty, false otherwise */
 	bool empty() const { return ll.empty(); }
 	/** Removes the list item at position.
